+++ mail/layouts/text.php
@@ -1,5 +1,6 @@
 <?php
 use yii\helpers\Html;
+use yii\helpers\Url;
 /** @var \yii\web\View $this view component instance */
 /** @var \yii\mail\MessageInterface $message the message being composed */
 /** @var string $content main view render result */
@@ -7,7 +8,6 @@ use yii\helpers\Html;
 
 <?php $this->beginPage() ?>
 <?php $this->beginBody() ?>
-<h1>Newsletters from <?= Yii::$app->getHomeUrl() ?></h1>
 <?= $content ?>
 <?php $this->endBody() ?>
 <?php $this->endPage() ?>