+++ mail/newsletters-text.php
@@ -1,0 +1,9 @@
+<?php
+
+use yii\helpers\Url;
+use wdmg\helpers\StringHelper;
+
+?>
+Newsletters from <?= Url::base(true) ?>
+=====================================
+<?= StringHelper::stripTags($content, "", " ") ?>