+++ views/list/index.php
@@ -4,6 +4,9 @@ use yii\helpers\Html;
 use yii\grid\GridView;
 use yii\widgets\Pjax;
 use wdmg\widgets\SelectInput;
+use wdmg\helpers\StringHelper;
+use yii\helpers\Url;
+use yii\bootstrap\Modal;
 
 /* @var $this yii\web\View */
 /* @var $searchModel wdmg\newsletters\models\NewslettersSearch */
@@ -29,12 +32,71 @@ $this->params['breadcrumbs'][] = $this->title;
             ['class' => 'yii\grid\SerialColumn'],
 
             'title',
-            'description',
+            /*'description',*/
             'subject',
-            'content',
-            /*'layouts',*/
-            'recipients',
-            /*'unique_token',*/
+            [
+                'attribute' => 'content',
+                'format' => 'text',
+                'value' => function($data) {
+                    return StringHelper::truncateWords(StringHelper::stripTags($data->content, "", " "),12,'…');
+                }
+            ],
+            [
+                'attribute' => 'recipients',
+                'format' => 'html',
+                'headerOptions' => [
+                    'class' => 'text-center'
+                ],
+                'contentOptions' => [
+                    'class' => 'text-center'
+                ],
+                'value' => function($data) use ($searchModel) {
+                    $html = '';
+                    $lists = [];
+                    $emails = [];
+                    $data = \yii\helpers\Json::decode($data->recipients);
+                    foreach ($data as $key => $item) {
+
+                        if (preg_match('/email_id:(\d)/', $key)) {
+                            $emails[] = '<span class="label label-info">' . $item . '</span>';
+                        } elseif (preg_match('/list_id:(\d)/', $key, $match)) {
+                            $count = $searchModel->getSubscribersCount(['list_id' => intval($match[1])]);
+                            if (!is_null($count))
+                                $lists[] = '<span class="label label-warning">' . $item . ' ('.$count.')</span>';
+                            else
+                                $lists[] = '<span class="label label-warning">' . $item .'</span>';
+
+                        } else {
+                            $emails[] = '<span class="label label-default">' . $item . '</span>';
+                        }
+                    }
+
+                    $onMore = false;
+                    if (count($lists) >= 2)
+                        $onMore = true;
+
+                    if (count($emails) >= 4)
+                        $onMore = true;
+
+                    if ($onMore)
+                        $html .= join(array_slice($lists, 0, 2), " ");
+                    else
+                        $html .= join($lists, " ");
+
+                    $html .= " ";
+
+                    if ($onMore)
+                        $html .= join(array_slice($emails, 0, 4), " ");
+                    else
+                        $html .= join($emails, " ");
+
+                    if ($onMore)
+                        return $html . "&nbsp;… ";
+                    else
+                        return $html;
+
+                }
+            ],
             [
                 'attribute' => 'status',
                 'format' => 'html',
@@ -70,7 +132,42 @@ $this->params['breadcrumbs'][] = $this->title;
                 'class' => 'yii\grid\ActionColumn',
                 'header' => Yii::t('app/modules/newsletters', 'Actions'),
                 'contentOptions' => [
-                    'class' => 'text-center'
+                    'class' => 'text-center',
+                    'style' => 'min-width:120px',
+                ],
+                'buttons'=> [
+                    'refresh' => function($url, $data, $key) {
+                        return Html::a('Refresh&nbsp;<span class="glyphicon glyphicon-refresh"></span>', Url::to(['list/process', 'newsletter' => 'refresh', 'id' => $data['id']]), [
+                            'title' => Yii::t('app/modules/newsletters', 'Refresh progress'),
+                            'data-toggle' => 'refresh-progress',
+                            'data-id' => $key,
+                            'data-pjax' => '0'
+                        ]);
+                    },
+                    'stop' => function($url, $data, $key) {
+                        return Html::a('Stop&nbsp;<span class="glyphicon glyphicon-stop"></span>', Url::to(['list/process', 'newsletter' => 'stop', 'id' => $data['id']]), [
+                            'title' => Yii::t('app/modules/newsletters', 'Stop newsletter'),
+                            'class' => 'text-danger',
+                            'data-id' => $key,
+                            'data-pjax' => '0'
+                        ]);
+                    },
+                    'pause' => function($url, $data, $key) {
+                        return Html::a('Suspend&nbsp;<span class="glyphicon glyphicon-pause"></span>', Url::to(['list/process', 'newsletter' => 'pause', 'id' => $data['id']]), [
+                            'title' => Yii::t('app/modules/newsletters', 'Suspend newsletter'),
+                            'class' => 'text-warning',
+                            'data-id' => $key,
+                            'data-pjax' => '0'
+                        ]);
+                    },
+                    'play' => function($url, $data, $key) {
+                        return Html::a('Start&nbsp;<span class="glyphicon glyphicon-play"></span>', Url::to(['list/process', 'newsletter' => 'run', 'id' => $data['id']]), [
+                            'title' => Yii::t('app/modules/newsletters', 'Start newsletter'),
+                            'class' => 'text-success',
+                            'data-id' => $key,
+                            'data-pjax' => '0'
+                        ]);
+                    },
                 ],
                 'urlCreator' => function ($action, $model, $key, $index) {
 
@@ -83,7 +180,8 @@ $this->params['breadcrumbs'][] = $this->title;
                     if ($action === 'delete')
                         return \yii\helpers\Url::toRoute(['list/delete', 'id' => $key]);
 
-                }
+                },
+                'template' => '{view} {update} {delete} {refresh} {play} {stop} {pause}'
             ],
         ],
         'pager' => [
@@ -102,7 +200,7 @@ $this->params['breadcrumbs'][] = $this->title;
             'nextPageLabel'  => Yii::t('app/modules/newsletters', 'Next page &rarr;')
         ],
     ]); ?>
-
+    <hr/>
     <div>
         <?= Html::a(Yii::t('app/modules/newsletters', 'Add newsletter'), ['list/create'], ['class' => 'btn btn-success pull-right']) ?>
     </div>